s3torchbenchmarking/pyproject.toml
[build-system]
requires = ["setuptools", "build"]
build-backend = "setuptools.build_meta"

[project]
name = "s3torchbenchmarking"
version = "0.0.1"
description = "Tools to run and compare benchmarks against various PyTorch connectors like the s3torchconnector."
requires-python = ">=3.8,<3.13"
readme = "README.md"
dependencies = [
    # TODO: remove "!= 2.3.0" restriction once https://github.com/pytorch/data/issues/1244 is fixed
    # TODO: remove "< 2.5" restriction once https://github.com/pytorch/pytorch/issues/138333 is fixed
    "torch >= 2.0.1, != 2.3.0, < 2.5",
    "lightning >= 2.0",
    "s3torchconnector",
    "hydra-core",
    "torchdata>=0.6.1",
    "torchvision",
    "s3fs>=2024.6.1",
    "transformers",
    "numpy",
    "psutil",
    "pynvml",
    "boto3",
    "prefixed",
    "click",
    "accelerate",
    "pandas",
    "requests",
]

[project.optional-dependencies]
test = [
    "pytest"
]

[project.scripts]
s3torch-benchmark = "s3torchbenchmarking.benchmark:run_experiment"
s3torch-datagen = "s3torchbenchmarking.datagen:synthesize_dataset"
